Add File Upload
Kurang Edit Gambar

+++ src/controllers/mahasiswaController.js
@@ -1,3 +1,5 @@
+const fs = require('fs');
+const path = require('path');
 const MahasiswaModel = require('../models/mahasiswaModel');
 
 exports.getAllMahasiswa = async (req, res) => {
@@ -10,7 +12,7 @@ exports.getAllMahasiswa = async (req, res) => {
 };
 
 exports.createMahasiswa = async (req, res) => {
-  const { NIM, Nama, email, alamat } = req.body;
+  const { NIM, Nama, email, alamat, imageName } = req.body;
 
   try {
     // Periksa apakah NIM sudah terdaftar sebelumnya
@@ -19,7 +21,7 @@ exports.createMahasiswa = async (req, res) => {
       return res.status(400).json({ message: 'NIM telah terdaftar' });
     }
 
-    const newMahasiswa = new MahasiswaModel({ NIM, Nama, email, alamat });
+    const newMahasiswa = new MahasiswaModel({ NIM, Nama, email, alamat, imageName });
     const savedMahasiswa = await newMahasiswa.save();
     res.status(201).json(savedMahasiswa);
   } catch (error) {
@@ -32,26 +34,55 @@ exports.deleteMahasiswa = async (req, res) => {
   const { id } = req.params;
 
   try {
+    const mahasiswa = await MahasiswaModel.findById(id);
+    if (!mahasiswa) {
+      return res.status(404).json({ message: 'Mahasiswa not found' });
+    }
+
+    // Delete the associated image file
+    if (mahasiswa.imageName) {
+      const imagePath = path.join(__dirname, '..', 'uploads/img', mahasiswa.imageName);
+      fs.unlinkSync(imagePath);
+    }
+
     await MahasiswaModel.findByIdAndDelete(id);
     res.status(204).send();
   } catch (error) {
+    console.error(error);
     res.status(500).json({ message: 'Error deleting mahasiswa' });
   }
 };
 
 exports.editMahasiswa = async (req, res) => {
   const { id } = req.params;
-  const { NIM, Nama, email, alamat } = req.body;
+  const { Nama, email, alamat, imageName } = req.body;
 
   try {
+    let updatedMahasiswa = { Nama, email, alamat };
 
-    const updatedMahasiswa = await MahasiswaModel.findByIdAndUpdate(
+    if (imageName) {
+      // If the imageName is provided, update it as well
+      updatedMahasiswa = { ...updatedMahasiswa, imageName };
+
+      // Find the existing Mahasiswa to get the old image name
+      const existingMahasiswa = await MahasiswaModel.findById(id);
+
+      // Delete the old image file if it exists
+      if (existingMahasiswa.imageName) {
+        const imagePath = path.join(__dirname, '..', 'uploads/img', existingMahasiswa.imageName);
+        fs.unlinkSync(imagePath); // Remove the file from disk
+      }
+    }
+
+    const updatedData = await MahasiswaModel.findByIdAndUpdate(
       id,
-      { Nama, email, alamat },
+      updatedMahasiswa,
       { new: true }
     );
-    res.json(updatedMahasiswa);
+
+    res.json(updatedData);
   } catch (error) {
     res.status(500).json({ message: 'Error editing mahasiswa' });
   }
 };
+
